backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.1
faiss-cpu==1.8.0.post1
pydantic==2.9.1
python-dateutil==2.9.0
python-multipart==0.0.9
